requirements.txt
wordcloud==1.8.2.2
pandas==1.3.5
dash==2.3.1
redis==4.3.4
kafka-python==2.0.2